pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-fluency"
version = "0.1.0"
description = "Edge-ready pronunciation/fluency classifier pipeline"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "pandas>=2.1",
    "pyyaml>=6.0",
    "torch>=2.2",
    "torchaudio>=2.2",
    "librosa>=0.10",
    "soundfile>=0.12",
    "pydub>=0.25",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "onnx>=1.15",
    "onnxruntime>=1.16",
    "psutil>=5.9",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "datasets>=2.16",
    "tqdm>=4.66",
    "rich>=13.7",
    "click>=8.1",
    "sounddevice>=0.4",
    "pytest>=7.4",
    "pytest-cov>=4.1"
]

[project.optional-dependencies]
device = ["adb-shell>=0.4"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
